+++ CHANGELOG.md
@@ -7,6 +7,13 @@ All notable changes to this project will be documented in this file.
 
 See [keep a changelog](https://keepachangelog.com/en/1.0.0/) for information about writing changes to this log.
 
+## [develop]
+
+### Added
+
+- Added User booking cache entity.
+- Added User booking cache service.
+
 ## [1.1.2] - 2023-08-29
 
 ### Changed